Add files via upload
In this project, i used the selenium library to scrape and to load the html page, because this web site use ajax. 

Firstly, i scraped the links of the announcements figured in the page, because after refreshing the page, the blocks that contain data change. Next, i access every sigle link to scrape all the informations that can be useful. And finally, i put all the data collected in one file : Announcements.csv.